mcp_servers/requirements.txt
mcp>=1.0.0
pyyaml>=6.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
ruff>=0.1.0
bandit>=1.7.0
